move tabs to common folder
