+++ components/pages/submission-system/program-clinical-submission/FilesNavigator/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import VerticalTabs from 'uikit/VerticalTabs';
+import VerticalTabs from 'uikit/Tabs/Vertical';
 import { css } from 'uikit';
 import Typography from 'uikit/Typography';
 import { ClinicalSubmissionEntityFile, ClinicalSubmissionQueryData } from '../types';
